LayoutTests/http/wpt/wk-web-animations/timing-model/keyframe-effect.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Check that the KeyframeEffect interface is defined and exposes a constructor</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

internals.pauseTimeline(document.timeline);

test(t => {
    assert_own_property(window, "KeyframeEffect");
}, "The KeyframeEffect interface is defined.");

test(t => {
    assert_throws(new TypeError, () => new KeyframeEffect);
    assert_throws(new TypeError, () => new KeyframeEffect(document.createElement("div")));
}, "The KeyframeEffect interface cannot be constructed without enough parameters.");

test(t => {
    assert_throws(new TypeError, () => new KeyframeEffect(0, null));
    assert_throws(new TypeError, () => new KeyframeEffect({}, null));
    assert_throws(new TypeError, () => new KeyframeEffect([], null));
    assert_throws(new TypeError, () => new KeyframeEffect(false, null));
    assert_throws(new TypeError, () => new KeyframeEffect('element', null));
}, "The KeyframeEffect interface constructor throws if the first parameter has the wrong type.");

test(t => {
    assert_equals((new KeyframeEffect(document.body, null)).target, document.body);
}, "The KeyframeEffect interface target is set to the provided Element.");

test(t => {
    const effect = new KeyframeEffect(document.body, null);
    assert_readonly(effect, "target");
}, "A KeyframeEffect object target is read-only.");

</script>
</body>
